gpuserver/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
websockets==13.1
pydantic==2.10.0
python-dotenv==1.0.1
pydantic-settings==2.6.0
httpx==0.27.0
python-multipart==0.0.21  # For file upload support
# LLM dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-ollama>=0.0.1
# ASR dependencies
openai-whisper
soundfile
# TTS dependencies
edge-tts
